backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
motor==3.3.2
pymongo==4.6.3
pydantic==2.6.4
email-validator==2.1.1
passlib==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
python-multipart==0.0.9
